Once user changes lights using this app, light automation ends for today
User changes the lights via voice command will set a pref flag that
records the last time lights have changed by a user. The light automation
will not adjust lights if this flag was set on the same day. This means
that if the user adjusts lights of the day, the app will not disrupt
the user's choice.

Automation is set back to normal if the user connects to wifi after 10min
being disconnected; mostly lightly showing that the user was out and
when they come back, the lights should re-adjust back.