Adjust last interaction time to avoid early morning changes
We substract the time by the late night time (hours and minutes) so that
the lights changes after midnight do not count as interacted time for
the day and it sets back to auto mode. We used to check the year, month
and day of month to track if interaction was the same day. Subtracting it
makes early morning changes to the previous day allowing automation to
occur later at sunset.